Custom_Services/pyproject.toml
[tool.poetry]
name = "nidcpower_source_dc_voltage"
version = "0.5.0"
package-mode = false
description = "Measurement plug-in example that sources and measures a DC voltage with an NI SMU."
authors = ["National Instruments"]

[tool.poetry.dependencies]
python = "^3.8"
nidcpower = { version = ">=1.4.4", extras = ["grpc"] }
ni-measurement-plugin-sdk-service = {version = "^2.0.0"}
click = ">=7.1.2, !=8.1.4" # mypy fails with click 8.1.4: https://github.com/pallets/click/issues/2558
grpcio = "*"
protobuf = "4.25.4"
grpcio-tools = "1.59.0"

[tool.poetry.group.dev.dependencies]
ni-python-styleguide = ">=0.4.1"
mypy = ">=1.0"
grpc-stubs = "^1.53"
# Uncomment to use prerelease dependencies.
# nidcpower = { git = "https://github.com/ni/nimi-python.git", subdirectory = "generated/nidcpower", extras = ["grpc"] }
# ni-measurement-plugin-sdk-service = {path = "../../packages/service", develop = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "hightime.*",
    "nidcpower.*",
]
ignore_missing_imports = true
